% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_import.R
\name{asana_import}
\alias{asana_import}
\title{Gets tasks from a project in Asana}
\usage{
asana_import(ASANA_ACCESS_TOKEN = Sys.getenv("ASANA_ACCESS_TOKEN"),
  project_gid = Sys.getenv("ASANA_MYTASKS_PROJECT_ID"),
  board_column = NA, section = NA)
}
\arguments{
\item{ASANA_ACCESS_TOKEN}{An Asana access token.
You can get your own personal Asana access token here:
https://asana.com/guide/help/api/api#gl-access-tokens}

\item{project_gid}{The gid identifier for a specific project in Asana,
found in the URL}

\item{board_column}{If importing from an Asana project in board view,
specify the column that you would like to import and sort}

\item{section}{If importing from an Asana project in list view,
specify a section that you would like to import and sort}
}
\description{
Gets tasks from a project in Asana
}
\details{
Follow the instructions for the asana package to make usage easier
https://github.com/datacamp/asana/blob/master/README.md
}
\examples{
asana_import(
ASANA_ACCESS_TOKEN = "add_your_34_chr_asana_access_token",
project_gid = "123456789101112")
}
