% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_move_task.R
\name{asana_move_task}
\alias{asana_move_task}
\title{Move A Task Within A Project}
\usage{
asana_move_task(task_to_move, insert_before = NULL, project,
  ASANA_ACCESS_TOKEN = Sys.getenv("ASANA_ACCESS_TOKEN"))
}
\arguments{
\item{task_to_move}{The gid of the task that you want to move}

\item{insert_before}{The gid of the task that you would like to use as reference for
task placement. Defaults to top of list}

\item{project}{The project gid}

\item{section}{An Asana access token.
You can get your own personal Asana access token here:
https://asana.com/guide/help/api/api#gl-access-tokens}
}
\description{
Move A Task Within A Project
}
\details{
Follow the instructions for the asana package to make usage easier
https://github.com/datacamp/asana/blob/master/README.md
}
\examples{
asana_move_task() # example coming soon
}
