% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asana_section_gids.R
\name{asana_section_gids}
\alias{asana_section_gids}
\title{Get A Tibble Of All Sections Within A Project And Their gids}
\usage{
asana_section_gids(project, section_str = NA,
  include_no_section = FALSE,
  ASANA_ACCESS_TOKEN = Sys.getenv("ASANA_ACCESS_TOKEN"))
}
\arguments{
\item{project}{The project gid}

\item{section_str}{A character string contained within the section name(s)
you'd like to return.}

\item{ASANA_ACCESS_TOKEN}{An Asana access token.
You can get your own personal Asana access token here:
https://asana.com/guide/help/api/api#gl-access-tokens}
}
\description{
Get A Tibble Of All Sections Within A Project And Their gids
}
\details{
Follow the instructions for the asana package to make usage easier
https://github.com/datacamp/asana/blob/master/README.md
}
\examples{
asana_section_gids() # example coming soon
}
