% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_performance_order.R
\name{bin_performance_order}
\alias{bin_performance_order}
\title{Get the order to do tasks based on Urgency, Importance, and Urgency Bias.}
\usage{
bin_performance_order(urgency_bias = 1.4, bin_u_max = 5,
  bin_i_max = 5, bin_u_min = 1, bin_i_min = 1)
}
\arguments{
\item{urgency_bias}{A bias toward weighting Urgency more heavily than
Importance. Urgency bias above 1 favors Urgency, below 1 favors importance.}

\item{bin_u_max}{The highest integer tank for Urgency}

\item{bin_i_max}{The highest integer tank for Importance}

\item{bin_u_min}{The lowest integer tank for Urgency}

\item{bin_i_min}{The lowest integer tank for Importance}
}
\description{
Each combination of Urgency and Importance can be imagined as a set of x and
y coordinates. A line can then be shifted from the point which represents
tasks that should be performed soonest to the point that represents tasks
that should be performed last. This finds the order in which that line
passes through those points

Returns a dataframe with three columns, the u rank, the i rank, and
performance rank. A lower number for performance rank means it should be
performed sooner.
}
\examples{
bin_performance_order(urgency_bias = 1.4)
}
