% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisection_insert.R
\name{bisection_insert}
\alias{bisection_insert}
\title{Insert a single task efficiently into an already-sorted list}
\usage{
bisection_insert(task, tasks, tasks_column = "Task")
}
\arguments{
\item{task}{The tast to be sorted, a character}

\item{tasks}{The pre-sorted list, a tibble}

\item{tasks_column}{The column name of the tasks within the presorted list}
}
\description{
Insert a single task efficiently into an already-sorted list
}
\examples{
task <- "Fix the roof"
tasks <- tibble("To-Do" = c("task a", "task b", "task c",
"task d", "task e", "task f", "task g"))
bisection_insert(task, tasks, tasks_column = "To-Do")
}
