% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sections.R
\name{build_sections}
\alias{build_sections}
\title{Create new sections in a project in Asana}
\usage{
build_sections(sections, project_gid)
}
\arguments{
\item{sections}{A vector of new section titles to be added to the
project, in order. If a section already exists it will not be duplicated.}

\item{project_gid}{The gid identifier for a specific project in Asana,}
}
\description{
Create new sections in a project in Asana
}
\examples{
project_gid <- "1234567891011121"
sections <- c("Do", "Delegate", "Schedule")
build_sections(sections, project_gid)

}
