% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Choose Which Of Two Options Is More Important}
\usage{
compare(choice1, choice2)
}
\arguments{
\item{choice1}{The first item, a character}

\item{choice2}{The second item, a character}
}
\description{
To be used within the quickSort() function from package rje
Output is either -1 or 1, similar output to greaterThan() from rje
}
\examples{
choice1 <- "Fix the roof"
choice2 <- "Clean the living room"
compare(choice1, choice2)
}
