% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decider.R
\name{decider}
\alias{decider}
\title{Decide What To Do And When To Do It}
\usage{
decider(input_type = "asana",
  asana_project_gid = Sys.getenv("ASANA_PROJECT_ID"),
  run_shiny = FALSE, csv_path, csv_task_column_name = "Task",
  testing_task_num = NA)
}
\arguments{
\item{input_type}{The method to import tasks into decider.
Can be "asana" or "csv"}

\item{asana_project_gid}{The global identifier for an asana project.
Can usually be found in the project URL.}

\item{run_shiny}{Choose whether to run the GUI for decider}

\item{csv_path}{The filepath to a csv with a list of tasks}

\item{csv_task_column_name}{The name of the column containing the task list
within a csv file}

\item{testing_task_num}{A number of tasks you'd like to sample from the
total list of tasks. For testing decider functionality.}
}
\description{
Decider dramatically reduces the amount of time and mental
effort it takes
to decide which tasks to do and in what order to do them
Primarily focused on integrating with Asana, but can be modified to work
with any csv or list of tasks

We accomplish this goal by combining together these established concepts:

The Eisenhower Matrix:
https://www.eisenhower.me/eisenhower-matrix/

Likert Scale:
https://www.simplypsychology.org/likert-scale.html

QuickSort:
https://www.youtube.com/watch?v=WaNLJf8xzC4
}
\details{
Follow the instructions for the asana package to make usage easier
https://github.com/datacamp/asana/blob/master/README.md
}
