% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_jointBLRM.R
\name{fit_jointBLRM}
\alias{fit_jointBLRM}
\title{Fit the joint BLRM and extract a stanfit object}
\usage{
fit_jointBLRM(
   dose1,
   dose2,
   dose.ref1,
   dose.ref2,
   n.pat,
   n.dlt,
   trial,
   MAP.prior = FALSE,
   saturating = FALSE,
   prior.mu = list(mu_a1 =  c(logit(0.33), 2),
                   mu_b1 =  c(0,          1),
                   mu_a2 =  c(logit(0.33), 2),
                   mu_b2 =  c(0,          1),
                   mu_eta = c(0,          1.121)),
   prior.tau = list(tau_a1 =  c(log(0.25),  log(2)/1.96),
                    tau_b1 =  c(log(0.125), log(2)/1.96),
                    tau_a2 =  c(log(0.25),  log(2)/1.96),
                    tau_b2 =  c(log(0.125), log(2)/1.96),
                    tau_eta = c(log(0.125), log(2)/1.96)),
   ...)
}
\arguments{
\item{dose1}{Numerical vector that lists the dose levels of the first compound applied to each
patient. The length must be equal to the number of cohorts. If the first compound was not
administered to a cohort, use 0 for this entry.}

\item{dose2}{Numerical vector that lists the dose levels of the second compound applied to each
patient. The length must be equal to the number of cohorts. If the second compound was not
administered to a cohort, use 0 for this entry.}

\item{dose.ref1}{Positive number, reference dose for compound 1.}

\item{dose.ref2}{Positive number, reference dose for compound 2.}

\item{n.pat}{Numerical vector that lists the number of patients for each cohort.}

\item{n.dlt}{Numerical vector that lists the number of DLTs for each cohort.}

\item{trial}{Numerical vector that indicates to which trial a cohort is assigned to.
The entries can either be numericals or strings, where \code{1} and \code{"1"}
would be interpreted as the same trial. The entries of \code{trial} are internally
converted to numbers from 1 to the number of different studies if not already
given in this form.}

\item{MAP.prior}{Optional logical, defaults to \code{FALSE}. If \code{TRUE}, the returned stanfit contains an additional
trial number that provides the resulting meta-analytic predictive prior for a new trial.}

\item{saturating}{Optional logical, defaults to \code{FALSE}. If \code{TRUE}, the BLRM will be using a saturating interaction term as described in
\code{\link[OncoBayes2:blrm_formula_saturating]{OncoBayes2::blrm_formula_saturating}()}. Also refer to the Details section in the documentation
of \code{\link[OncoBLRM:scenario_jointBLRM]{scenario_jointBLRM}()}.}

\item{prior.mu}{Optional list that specifies the prior.
Same format as documented in \code{\link[OncoBLRM:scenario_jointBLRM]{scenario_jointBLRM}()}.}

\item{prior.tau}{Optional list that specifies the prior.
Same format as documented in \code{\link[OncoBLRM:scenario_jointBLRM]{scenario_jointBLRM}()}.}

\item{...}{Additional optional arguments that are passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling}()}. These
include the settings for MCMC and the list of parameters which are returned in the stanfit object. Refer
to the \code{\link[rstan:rstan]{rstan-package}} for an overview.}
}
\value{
An object of class \code{\link[rstan:stanfit-class]{rstan::stanfit-class}} from the
\code{\link[rstan:rstan]{rstan-package}}.
}
\description{
Call the \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling}()} method for
the joint BLRM to extract a \code{\link[rstan:stanfit-class]{rstan::stanfit-class}} object.
Refer to \code{\link[OncoBLRM:scenario_jointBLRM]{scenario_jointBLRM}()} for a
documentation of the underlying model.
}
\details{
Refer to \code{\link[OncoBLRM:scenario_jointBLRM]{scenario_jointBLRM}()} for
a description of the underlying model.
}
\examples{
\dontrun{
  fit <- fit_jointBLRM(
              dose1 = c(1, 2, 4, 8, 12, 0,  0,  0,  2,  4,  6),
              dose2 = c(0, 0, 0, 0, 0,  10, 20, 30, 10, 10, 20),
              dose.ref1 = 12,
              dose.ref2 = 30,
              n.pat = c(3, 3, 3, 3, 9,  3,  3,  6,  3,  3,  3),
              n.dlt = c(0, 0, 0, 0, 2,  0,  0,  1,  0,  0,  1),
              trial = c(1, 1, 1, 1, 1,  2,  2,  2,  3,  3,  3),

              iter = 10000,
              chains = 4,
              control = list(adapt_delta = 0.95),
              pars = c("log_ab")
 )

 #print posterior summary
 print(fit)

 #posterior density plots for parameters of trial 3.
 plot <- rstan::stan_plot(fit, pars=paste0("log_ab[3,", 1:5, "]"),
                          show_density=TRUE)
 print(plot)

 #extract samples in matrix format
 samples <- as.matrix(fit, pars="log_ab")
 head(samples)
}
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. \url{https://mc-stan.org}.

Neuenschwander, B., Branson, M., & Gsponer, T. (2008). Critical aspects of the Bayesian approach to phase I cancer trials.
Statistics in medicine, 27(13), 2420-2439, \url{doi:10.1002/sim.3230}.

Neuenschwander, B., Matano, A., Tang, Z., Roychoudhury, S., Wandel, S., & Bailey, S. (2014). A Bayesian Industry Approach to Phase I Combination Trials in Oncology.
In: Zhao. W & Yang, H. (editors). Statistical methods in drug combination studies. Chapman and Hall/CRC, 95-135, \url{doi:10.1201/b17965}.

Neuenschwander, B., Roychoudhury, S., & Schmidli, H. (2016). On the use of co-data in clinical trials.
Statistics in Biopharmaceutical Research, 8(3), 345-354, \url{doi:10.1080/19466315.2016.1174149}.
}
\seealso{
\code{\link[OncoBLRM:scenario_jointBLRM]{scenario_jointBLRM}()},
\code{\link[OncoBLRM:get_pred_tox]{get_pred_tox}()}, \code{\link[OncoBLRM:get_MAP]{get_MAP}()},
\code{\link[rstan:stanmodel-method-sampling]{rstan::sampling}()},
\code{\link[rstan:stanfit-class]{rstan::stanfit-class}}, \code{\link[rstan:rstan]{rstan-package}}.
}
