% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pred_tox.R
\name{get_pred_tox}
\alias{get_pred_tox}
\alias{get_pred_tox_mono}
\title{Compute predictive DLT rates for a new trial from historical data}
\usage{
get_pred_tox(
 doses.of.interest,
 dose1,
 dose2,
 dose.ref1,
 dose.ref2,
 n.pat,
 n.dlt,
 trial,
 prior.mu = list(mu_a1 =  c(logit(0.33), 2),
                 mu_b1 =  c(0,           1),
                 mu_a2 =  c(logit(0.33), 2),
                 mu_b2 =  c(0,           1),
                 mu_eta = c(0,           1.121)
                 ),
 prior.tau = list(tau_a1 =  c(log(0.25),  log(2)/1.96),
                  tau_b1 =  c(log(0.125), log(2)/1.96),
                  tau_a2 =  c(log(0.25),  log(2)/1.96),
                  tau_b2 =  c(log(0.125), log(2)/1.96),
                  tau_eta = c(log(0.125), log(2)/1.96)
                  ),
 saturating = FALSE,
 return.samples = FALSE,
 probs = c(0.025, 0.5, 0.975),
 dosing.intervals = c(0.16, 0.33),
 ...
)

get_pred_tox_mono(
 doses.of.interest,
 dose,
 dose.ref,
 n.pat,
 n.dlt,
 trial,
 prior.mu.mono = list(mu_a =  c(logit(0.33), 2),
                      mu_b =  c(0,          1)),
 prior.tau.mono = list(tau_a =  c(log(0.25),  log(2)/1.96),
                       tau_b =  c(log(0.125), log(2)/1.96)),
 return.samples = FALSE,
 probs = c(0.025, 0.5, 0.975),
 dosing.intervals = c(0.16, 0.33),
 ...
)
}
\arguments{
\item{doses.of.interest}{Numerical that defines the set of dose levels for
which the meta-analytic predictive DLT rate is to be computed. Its specification
differs between \code{get_pred_tox()} and \code{get_pred_tox_mono()}.

For \code{get_pred_tox()}: Can be either
a vector of length 2 that describes a single dose combination (where one of
the doses can be set to 0 for monotherapy doses), or a matrix with 2 rows in
which each column defines a dose combination.

For \code{get_pred_tox_mono()}: Can be either a vector giving a set of monotherapy
dose levels, or a matrix with exactly one row or exactly one column. In both cases,
each number in the argument is interpreted as a single dose.}

\item{dose1}{Numeric vector with non-negative entries. Describes the dose of
compound 1 administed to the patients of a cohort in the data, where cohort \code{i}
is assumed to have received \code{dose1[i]} of compound 1. Note that \code{dose1[i]}
can be set to 0 to indicate that a cohort received monotherapy with compound 2.}

\item{dose2}{Numeric vector with non-negative entries. Describes the dose of
compound 2 administed to the patients of a cohort in the data, where cohort \code{i}
is assumed to have received \code{dose2[i]} of compound 2. Note that \code{dose2[i]}
can be set to 0 to indicate that a cohort received monotherapy with compound 1.}

\item{dose.ref1}{Positive number, reference dose for compound 1.}

\item{dose.ref2}{Positive number, reference dose for compound 2.}

\item{n.pat}{Numeric vector with non-negative whole numbers as entries. \code{n.pat[i]}
indicates the number of patients that were treated in cohort \code{i}.}

\item{n.dlt}{Numeric vector with non-negative whole numbers as entries. \code{n.dlt[i]}
indicates the number of patients that experienced DLT in cohort \code{i}.}

\item{trial}{Numerical or character vector that indicates to which trial a cohort is assigned to.
The entries can either be numericals or strings, where \code{1} and \code{"1"}
would be interpreted as the same trial. Note that the given entries are internally
converted to numbers from 1 to the number of different studies if not already
given in this form.}

\item{prior.mu}{Named list that specifies the distribution of the
hypermeans of the parameters of the combination therapy BLRM. Same format,
default, and requirements as documented for the function
\code{\link[decider:scenario_jointBLRM]{scenario_jointBLRM}()}.}

\item{prior.tau}{Named list that specifies the distribution of the
between-trial heterogeneities of the parameters of the combination therapy BLRM.
Same format, default, and requirements as documented for the function
\code{\link[decider:scenario_jointBLRM]{scenario_jointBLRM}()}.}

\item{saturating}{Optional logical, defaults to \code{FALSE}. If \code{TRUE}, the BLRM will be using a saturating interaction term as described in
\code{\link[OncoBayes2:blrm_formula_saturating]{OncoBayes2::blrm_formula_saturating}()}. Also refer to the Details section in the documentation
of \code{\link[decider:scenario_jointBLRM]{scenario_jointBLRM}()}.}

\item{return.samples}{Optional logical, defaults to \code{FALSE}. If \code{TRUE},
the function will return a matrix with samples from the MAP prior additionally
to a prior summary.}

\item{probs}{Optional numerical, defaults to \code{c(0.025, 0.5,  0.975)}.
Indicates which quantiles should be included in the posterior summaries.}

\item{dosing.intervals}{Optional numeric ascending positive entries between 0 and 1.
Defines the limits of the dosing intervals to be evaluated, similarly to
\code{\link[decider:scenario_jointBLRM]{scenario_jointBLRM}()}. The default value
is \code{c(0.16, 0.33)}.}

\item{...}{Optional additional arguments that are passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling}()}. These
include the settings for MCMC and the random seed. Refer
to the \code{\link[rstan:rstan]{rstan-package}} for an overview. Note that
the function will automatically use \code{iter=10000} and a \code{control} argument
containing \code{adapt_delta=0.9} to ensure a relatively good fit if the
arguments \code{iter} and \code{control} for \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling}()}
are not provided manually in \code{...}.}

\item{dose}{Numeric vector with non-negative entries that defines the monotherapy dose
levels administered to each cohort.}

\item{dose.ref}{Positive numerical, reference dose for monotherapy.}

\item{prior.mu.mono}{Optional named list that specifies the prior. Must have the
entries \code{mu_a} and \code{mu_b}, both of which must be vectors of length 2,
where the second entry is a positive number.\cr
The entries need to be:
\itemize{
\item{\code{...$mu_a}:\cr
Defaults to \code{c(logit(0.33), 2)}, provides mean and SD of the normal distribution used as
hypermean of \eqn{log(\alpha)}.}
\item{\code{...$mu_b}:\cr
Defaults to \code{c(0, 1)}, provides mean and SD of the normal distribution used as
hypermean of \eqn{log(\beta)}.}
}}

\item{prior.tau.mono}{Optional named list that specifies the prior. Must have the
entries \code{tau_a} and \code{tau_b}, both of which must be vectors of length 2,
where the second entry is a positive number.\cr
The entries need to be:
\itemize{
\item{\code{...$tau_a}:\cr
Defaults to \code{c(log(0.25), log(2)/1.96)}, provides mean and SD (on log-scale) of the log-normal distribution used as
between-trial heterogeneity of \eqn{log(\alpha)}.}
\item{\code{...$tau_b}:\cr
Defaults to \code{c(log(0.125), log(2)/1.96)}, provides mean and SD (on log-scale) of the log-normal distribution used as
between-trial heterogeneity of \eqn{log(\beta)}.}
}}
}
\description{
The function \code{get_pred_tox()} can be used to derive predictive
distributions of the DLT rates of a set of dose levels considered in a new
dose finding trial based on historical data from one or more dose-finding trials
based on a joint BLRM. The function supports both monotherapy and two-drug
combination therapy and is essentially a wrapper for \code{\link[decider:get_MAP]{get_MAP}()}
but additionally computes samples from the predictive DLT rates from the
drawn samples from the MAP prior. Note that this is equivalent to the posterior
distribution of the DLT rates of a trial that has not recorded
data yet from a joint BLRM including the specified co-data from other trials.

The function \code{get_pred_tox()} can be used for both monotherapy and
combination therapy.
For the former, the arguments \code{dose2} and \code{dose.ref2} can
be left out. The function \code{get_pred_tox_mono()} uses a slightly different
syntax and output structure that is specifically designed for settings consisting purely of
monotherapy trials.
}
\details{
Refer to the documentation of \code{\link[decider:scenario_jointBLRM]{scenario_jointBLRM}()}
for a description of the underlying model used to compute predictive distributions
of the DLT rates.

The function \code{get_pred_tox()} can optionally return the samples from the
predictive distribution, which can in turn be used to perform single-dose analyses
using a binomial-beta-mixture model. Refer to the documentation of \code{\link[decider:binomial_betamix]{binomial_betamix}()}
and the vignette on single dose confirmation for details.
}
\examples{
\dontrun{
#derive the predictive distribution of the DLT rate of the dose 10+24
#from historical data of three previous trials
pred <- get_pred_tox(
 doses.of.interest = c(10, 24),
 dose1 = c(2, 4, 8, 10, 14, 0,  0,  0,  4,  8,  10, 12),
 dose2 = c(0, 0, 0, 0,  0,  12, 24, 36, 24, 24, 24, 24),
 n.pat = c(3, 3, 3, 9,  6,  3,  12, 3,  3,  3,  9,  3),
 n.dlt = c(0, 0, 0, 0,  2,  0,  1,  0,  0,  0,  1,  1),
 trial = c(1, 1, 1, 1,  1,  2,  2,  2,  3,  3,  3,  3),
 dose.ref1 = 14,
 dose.ref2 = 24,
 return.samples = TRUE
)

#summary
pred$summary

#first eight samples from predictive DLT rate
pred$samples[, 1:8]

}
\dontrun{
#derive the predictive distribution of the DLT rates of the doses 8, 12, and
#14 (monotherapy) from historical data of two previous trials
pred_mono <- get_pred_tox_mono(
 doses.of.interest = c(8, 12, 14),
 dose  = c(2, 4, 8, 10, 14, 10, 14),
 n.pat = c(3, 3, 3, 9,  6,  6,  12),
 n.dlt = c(0, 0, 0, 0,  2,  0,  1),
 trial = c(1, 1, 1, 1,  1,  2,  2),
 dose.ref = 14,
 return.samples = TRUE
)

#summary
pred_mono$summary

#first eight samples from predictive DLT rates
pred_mono$samples[, 1:8]

}
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. \url{https://mc-stan.org}.

Neuenschwander, B., Branson, M., & Gsponer, T. (2008). Critical aspects of the Bayesian approach to phase I cancer trials.
Statistics in medicine, 27(13), 2420-2439, \url{doi:10.1002/sim.3230}.

Schmidli, H., Gsteiger, S., Roychoudhury, S., O'Hagan, A., Spiegelhalter, D., & Neuenschwander B. (2014).
Robust meta-analytic-predictive priors in clinical trials with historical control information.

Neuenschwander, B., Matano, A., Tang, Z., Roychoudhury, S., Wandel, S., & Bailey, S. (2014). A Bayesian Industry Approach to Phase I Combination Trials in Oncology.
In: Zhao. W & Yang, H. (editors). Statistical methods in drug combination studies. Chapman and Hall/CRC, 95-135, \url{doi:10.1201/b17965}.

Neuenschwander, B., Roychoudhury, S., & Schmidli, H. (2016). On the use of co-data in clinical trials.
Statistics in Biopharmaceutical Research, 8(3), 345-354, \url{doi:10.1080/19466315.2016.1174149}.
}
\seealso{
\code{\link[decider:get_MAP]{get_MAP}()},
\code{\link[decider:scenario_jointBLRM]{scenario_jointBLRM}()},
\code{\link[rstan:stanmodel-method-sampling]{rstan::sampling}()}, \code{\link[rstan:rstan]{rstan-package}}.
}
