% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_task.R
\name{place_task}
\alias{place_task}
\title{Move tasks to the proper section within a project}
\usage{
place_task(task, project_gid, section)
}
\arguments{
\item{task}{a one-row tibble of a task to be placed, contains Task and gid}

\item{project_gid}{The gid identifier for a specific project in Asana}

\item{section}{The section to move the task to}
}
\description{
Move tasks to the proper section within a project
}
\examples{
project_gid <- "1234567891011121"
sections <- c("Do", "Delegate", "Schedule")
build_sections(sections, project_gid)

}
