% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_covariate_jointBLRM.R
\name{scenario_list_covariate_jointBLRM}
\alias{scenario_list_covariate_jointBLRM}
\title{Process a list of data scenarios including a binary covariate in parallel}
\usage{
scenario_list_covariate_jointBLRM(data.list, file.names = NULL, ...)
}
\arguments{
\item{data.list}{List of hypothetical data scenarios. Each entry is passed to
\code{\link[decider:scenario_jointBLRM]{scenario_jointBLRM}()}.}

\item{file.names}{Optional character vector, defaults to \code{NULL}. The argument
has only an effect when results are saved to disk, i.e. when an additional argument
\code{path} is supplied that states the output location.
If a \code{path} is supplied and \code{file.names} is \code{NULL}, the \code{file.name} of
the output for scenario \eqn{i} will be \code{"scenario_i"}.
Otherwise, a vector with the same length as the \code{data.list} can be supplied as \code{file.names},
which is used as a file name for each scenario.}

\item{...}{Arguments that are passed to \code{\link[decider:scenario_jointBLRM]{scenario_jointBLRM}()},
e.g. priors, number of MCMC iterations, and number of chains. Can contain any arguments of
\code{scenario_jointBLRM()}, except for the argument \code{data}, which is taken from the \code{data.list} instead,
and for the argument \code{file.name}, which is taken from \code{file.names} instead.
Note in particular that \code{historical.data} can still be supplied, which is then included in every
evaluated scenario, additionally to the scenario-specific data from the \code{data.list} entries.}
}
\value{
Numbered list in which each entry contains the output of \code{\link[decider:scenario_jointBLRM]{scenario_jointBLRM}()}
for the corresponding trial in the \code{data.list}.
}
\description{
Wrapper for calling \code{\link[decider:scenario_covariate_jointBLRM]{scenario_covariate_jointBLRM}()}
for a list of data scenarios, and processing them in parallel. This function is mostly included
for convenience, e.g. to evaluate a number of scenarios that are used for prior adjustment
in a quick manner.

Please note that for parallel execution, a parallel backend needs to be registered by the user.
A basic example of this is available in the following vignette:
\code{vignette("parallelization_jointBLRM", package = "decider")}.

Be aware that input checks are performed by \code{\link[decider:scenario_jointBLRM]{scenario_jointBLRM}()}.
If this results in an error for one or more scenarios, the error is returned in the output list.
}
\details{
Be aware that parallelization requires the initialization of a parallel backend by the user.
Otherwise, the function is executed sequentially.
}
\seealso{
\code{\link[decider:scenario_covariate_jointBLRM]{scenario_covariate_jointBLRM}()},\code{\link[decider:scenario_jointBLRM]{scenario_jointBLRM}()},
\code{\link[rstan:stanmodel-method-sampling]{rstan::sampling}()},
\code{\link[rstan:rstan]{rstan-package}},
\code{\link[ggplot2:ggplot]{ggplot2::ggplot}}, \code{\link[ggplot2:ggplot2-package]{ggplot2-package}}.
}
