% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_jointBLRM.R
\name{scenario_list_jointBLRM}
\alias{scenario_list_jointBLRM}
\title{Process a list of data scenarios in parallel.}
\usage{
scenario_list_jointBLRM(data.list, n.cores = 1, file.names = NULL, ...)
}
\arguments{
\item{data.list}{List of hypothetical data scenarios. Each entry is passed to
\code{\link[OncoBLRM:scenario_jointBLRM]{scenario_jointBLRM}()}.}

\item{n.cores}{Optional integer, defaults to \code{1}. Indicates the number of workers
to be used during parallel processing. A value of \code{1} causes the scenarios to
be processed sequentially.}

\item{file.names}{Optional character vector, defaults to \code{NULL}. The argument
has only an effect when results are saved to disk, i.e. when an additional argument
\code{path} is supplied that states the output location.
If a \code{path} is supplied and \code{file.names} is \code{NULL}, the \code{file.name} of
the output for scenario \eqn{i} will be \code{"scenario_i"}.
Otherwise, a vector with the same length as the \code{data.list} can be supplied as \code{file.names},
which is used as a file name for each scenario.}

\item{...}{Arguments that are passed to \code{\link[OncoBLRM:scenario_jointBLRM]{scenario_jointBLRM}()},
e.g. priors, number of MCMC iterations, and number of chains. Can contain any arguments of
\code{scenario_jointBLRM()}, except for the argument \code{data}, which is taken from the \code{data.list} instead,
and for the argument \code{file.name}, which is taken from \code{file.names} instead.
Note in particular that \code{historical.data} can still be supplied, which is then included in every
evaluated scenario, additionally to the scenario-specific data from the \code{data.list} entries.}
}
\value{
Numbered list in which each entry contains the output of \code{\link[OncoBLRM:scenario_jointBLRM]{scenario_jointBLRM}()}
for the corresponding trial in the \code{data.list}.
}
\description{
Wrapper for calling \code{\link[OncoBLRM:scenario_jointBLRM]{scenario_jointBLRM}()}
for a list of data scenarios, and processing them in parallel. This function is mostly included
for convenience, e.g. to evaluate a number of scenarios that are used for prior adjustment
in a quick manner.

Be aware that input checks are performed by \code{\link[OncoBLRM:scenario_jointBLRM]{scenario_jointBLRM}()}.
If this results in an error for one or more scenarios, the error is returned in the output list.
}
\details{
Be aware that parallelization could also be performed at level of the chains
in Stan. However, we recommend to parallelize at level of the scenarios.
This way, also the generation of e.g. the plots and posterior toxicities are parallelized,
and, additionally, the initialization of the parallel backend needs to occur only once.
}
\seealso{
\code{\link[OncoBLRM:scenario_jointBLRM]{scenario_jointBLRM}()}, \code{\link[rstan:sampling]{rstan::sampling}()},
\code{\link[rstan:rstan-package]{rstan-package}},
\code{\link[ggplot2:ggplot]{ggplot2::ggplot}}, \code{\link[ggplot2:ggplot2-package]{ggplot2-package}}.
}
